+++ paymentsheet-example/src/androidTest/java/com/stripe/android/lpm/TestEmbedded.kt
@@ -21,6 +21,7 @@ import com.stripe.android.paymentsheet.example.playground.settings.SupportedPaym
 import com.stripe.android.test.core.FieldPopulator
 import com.stripe.android.test.core.TestParameters
 import com.stripe.android.utils.ForceNativeBankFlowTestRule
+import org.junit.Ignore
 import org.junit.Rule
 import org.junit.Test
 
@@ -85,6 +86,8 @@ internal class TestEmbedded : BasePlaygroundTest() {
         )
     }
 
+    // TODO(alexzhu): Enable after institutions are renamed (ir-rider-prefix)
+    @Ignore("Disabling test while institution name changes modifications are complete")
     @Test
     fun testUsBankAccount() {
         testDriver.confirmEmbeddedUsBankAccount(