+++ paymentsheet/src/main/java/com/stripe/android/lpmfoundations/paymentmethod/PaymentMethodMetadata.kt
@@ -81,6 +81,7 @@ internal data class PaymentMethodMetadata(
     val onBehalfOf: String?,
     val integrationMetadata: IntegrationMetadata,
     val analyticsMetadata: AnalyticsMetadata,
+    val experimentsData: ElementsSession.ExperimentsData?,
     val isTapToAddSupported: Boolean,
 ) : Parcelable {
 
@@ -375,6 +376,7 @@ internal data class PaymentMethodMetadata(
                 onBehalfOf = elementsSession.onBehalfOf,
                 integrationMetadata = integrationMetadata,
                 analyticsMetadata = analyticsMetadata,
+                experimentsData = elementsSession.experimentsData,
                 isTapToAddSupported = isTapToAddSupported,
             )
         }
@@ -440,6 +442,7 @@ internal data class PaymentMethodMetadata(
                 integrationMetadata = integrationMetadata,
                 analyticsMetadata = AnalyticsMetadata(emptyMap()), // This is unused in customer sheet.
                 isTapToAddSupported = false, // This is unused in customer sheet.
+                experimentsData = elementsSession.experimentsData,
             )
         }
     }